refactor(task-runner): extract TaskCard component for reuse
- Extract TaskCard with full Plan/Approve/Execute workflow from task-board
- Export component via package.json for cross-plugin usage
- Update site-builder task-panel to reuse TaskCard from task-runner
- Remove unused useEffect import from task-board
